% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ss}
\alias{get_ss}
\title{Get sufficient statistics}
\usage{
get_ss(dat)
}
\arguments{
\item{dat}{Data set with items as columns and examinees as rows. Missing
responses should be coded as NA.}
}
\value{
Vector of sufficient statistics
}
\description{
Find sufficient statistics from a data set
}
